Address reviewer's comment
